Adenda-ify the project, from `com.github.kliewkliew.salad` to `com.adendamedia.salad`
