+++ src/main/scala/com/adendamedia/salad/serde/Serde.scala
@@ -1,0 +1,11 @@
+package com.adendamedia.salad.serde
+
+/**
+  * A serializer-deserializer interface.
+  * @tparam D The decoded type.
+  * @tparam E The encoded type.
+  */
+trait Serde[D,E] {
+  def serialize(a: D): E
+  def deserialize(b: E): D
+}
