--- src/main/scala/com/github/kliewkliew/salad/serde/Serde.scala
@@ -1,11 +1,0 @@
-package com.github.kliewkliew.salad.serde
-
-/**
-  * A serializer-deserializer interface.
-  * @tparam D The decoded type.
-  * @tparam E The encoded type.
-  */
-trait Serde[D,E] {
-  def serialize(a: D): E
-  def deserialize(b: E): D
-}
